+++ demos/WebCam/Main.py
@@ -1,0 +1,5 @@
+from MiJuego import MiJuego
+
+# -- show time
+game = MiJuego()
+game.Run()